Move logic to ProfileController, add password changed confirmation email
