+++ resources/views/emails/password/password-changed.blade.php
@@ -1,0 +1,2 @@
+<p> Hey, you have successfully changed your password.</p>
+
